this version is more scalebale
in this version i broke most of dependancies into python packages
