+++ .idea/.name
@@ -1,1 +1,1 @@
-app.py
+lara_games_app.py
