+++ templates/base.html
@@ -47,11 +47,11 @@ th, td, tr {
         <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarToggle" aria-controls="navbarToggle" aria-expanded="True" aria-label="Toggle navigation">
           <span class="navbar-toggler-icon">
                  {% if current_user.is_authenticated %}
-              <a class="nav-item nav-link" href="{{ url_for('account') }}">Account</a>
-              <a class="nav-item nav-link" href="{{ url_for('logout') }}">Logout</a>
+              <a class="nav-item nav-link" href="{{ url_for('app.account') }}">Account</a>
+              <a class="nav-item nav-link" href="{{ url_for('app.logout') }}">Logout</a>
             {% else %}
-              <a class="nav-item nav-link" href="{{ url_for('login') }}">Login</a>
-              <a class="nav-item nav-link" href="{{ url_for('register') }}">Register</a>
+              <a class="nav-item nav-link" href="{{ url_for('app.login') }}">Login</a>
+              <a class="nav-item nav-link" href="{{ url_for('app.register') }}">Register</a>
             {% endif %}
           </span>
         </button>
@@ -65,11 +65,11 @@ th, td, tr {
           <!-- Navbar Right Side -->
           <div class=" col-sm-8 navbar-nav">
             {% if current_user.is_authenticated %}
-              <a class="nav-item nav-link" href="{{ url_for('account') }}">Account</a>
-              <a class="nav-item nav-link" href="{{ url_for('logout') }}">Logout</a>
+              <a class="nav-item nav-link" href="{{ url_for('app.account') }}">Account</a>
+              <a class="nav-item nav-link" href="{{ url_for('app.logout') }}">Logout</a>
             {% else %}
-              <a class="nav-item nav-link" href="{{ url_for('login') }}">Login</a>
-              <a class="nav-item nav-link" href="{{ url_for('register') }}">Register</a>
+              <a class="nav-item nav-link" href="{{ url_for('app.login') }}">Login</a>
+              <a class="nav-item nav-link" href="{{ url_for('app.register') }}">Register</a>
             {% endif %}
           </div>
         </div>
